# Improve performance of dynamic mmap flags counting