--- a/lib/segment/src/vector_storage/dense/dynamic_mmap_flags.rs
+++ b/lib/segment/src/vector_storage/dense/dynamic_mmap_flags.rs
@@ -331,6 +331,37 @@ mod tests {
     }
 
     #[test]
+    fn test_bitflags_counting() {
+        let dir = Builder::new().prefix("storage_dir").tempdir().unwrap();
+        let num_flags = 5003; // Prime number, not byte aligned
+        let mut rng = StdRng::seed_from_u64(42);
+
+        // Create randomized dynamic mmap flags to test counting
+        let mut dynamic_flags = DynamicMmapFlags::open(dir.path()).unwrap();
+        dynamic_flags.set_len(num_flags).unwrap();
+        let random_flags: Vec<bool> = iter::repeat_with(|| rng.gen()).take(num_flags).collect();
+        random_flags
+            .iter()
+            .enumerate()
+            .filter(|(_, flag)| **flag)
+            .for_each(|(i, _)| assert!(!dynamic_flags.set(i, true)));
+        dynamic_flags.flusher()().unwrap();
+
+        // Test count flags method
+        let count = dynamic_flags.count_flags().unwrap();
+
+        // Compare against manually counting every flag
+        let mut manual_count = 0;
+        for i in 0..num_flags {
+            if dynamic_flags.get(i) {
+                manual_count += 1;
+            }
+        }
+
+        assert_eq!(count, manual_count);
+    }
+
+    #[test]
     fn test_capacity() {
         assert_eq!(mmap_capacity_bytes(0), 128);
         assert_eq!(mmap_capacity_bytes(1), 128);