Improve our logic for counting deleted flags when opening the mmap based deleted flags from disk.

We're seeing some slowdowns when loading Qdrant collections, especially with memory mapping. I started this effort to improve one of the aspects of it.

The improvement is quite visible on large flag counts, such as 50m, where I see a 50x improvement. On low counts, such as 50k, things seem to be the other way around. I'm a bit confused on why this is the case, and may very well be noise from the benchmark itself. I think this potential slowdown like this is acceptable on low counts when including the potential increase on large counts.

Here are the bench results on various flag counts. Each second entry is the improved logic in this PR:

```bash
# 50_000 flags
dynamic-mmap-delete-count/manual-count-loop
                        time:   [257.00 µs 258.00 µs 259.20 µs]
dynamic-mmap-delete-count/count-ones
                        time:   [1.9390 ms 1.9463 ms 1.9541 ms]

# 1_000_000 flags
dynamic-mmap-delete-count/manual-count-loop
                        time:   [5.0834 ms 5.0928 ms 5.1028 ms]
dynamic-mmap-delete-count/count-ones
                        time:   [1.7795 ms 1.7846 ms 1.7939 ms]

# 50_000_000 flags
dynamic-mmap-delete-count/manual-count-loop
                        time:   [250.86 ms 251.92 ms 252.97 ms]
dynamic-mmap-delete-count/count-ones
                        time:   [5.2382 ms 5.2851 ms 5.3362 ms]

```

### All Submissions:

* [x] Contributions should target the `dev` branch. Did you create your branch from `dev`?
* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../../pulls) for the same update/change?

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### New Feature Submissions:

1. [x] Does your submission pass tests?
2. [x] Have you formatted your code locally using `cargo +nightly fmt --all` command prior to submission?
3. [x] Have you checked your code using `cargo clippy --all --all-features` command?